core/requirements.txt
# Core Django & essential libraries
Django==4.2.15
pillow==10.4.0
djangorestframework==3.15.2
python-decouple==3.8

# Third-party tools for API and filtering
django-filter==24.3
Markdown==3.7
drf-yasg[validation]
djangorestframework_simplejwt

# Deployment-related tools


# Code quality, formatting, and testing
black
flake8
flake8-django
